carpetas en templates, control de acceso en controllers, spring security

+++ src/main/java/com/if7100/controller/NivelEducativoController.java
@@ -1,25 +1,53 @@
 package com.if7100.controller;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
-import com.if7100.entity.NivelEducativo;
 
+import com.if7100.entity.Hecho;
+import com.if7100.entity.NivelEducativo;
+import com.if7100.entity.Perfil;
+import com.if7100.repository.UsuarioRepository;
 import com.if7100.service.NivelEducativoService;
+import com.if7100.service.PerfilService;
 
 
 @Controller
 public class NivelEducativoController {
 private NivelEducativoService nivelEducativoService;
+//instancias para control de acceso
+private UsuarioRepository usuarioRepository;
+private Perfil perfil;
+private PerfilService perfilService;
 	
-	public NivelEducativoController(NivelEducativoService nivelEducativoService) {
+	public NivelEducativoController(NivelEducativoService nivelEducativoService, PerfilService perfilService, UsuarioRepository usuarioRepository) {
 	
 		super();
 		this.nivelEducativoService = nivelEducativoService;
+		this.perfilService = perfilService;
+        this.usuarioRepository = usuarioRepository;
 	
 }
+	
+private void validarPerfil() {
+    	
+		try {
+			
+			Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+		    String username = authentication.getName();
+			
+			this.perfil = new Perfil(perfilService.getPerfilById(usuarioRepository.findByCVCedula(username).getCIPerfil()));
+			
+		}catch (Exception e) {
+			// TODO: handle exception
+		}
+		
+	}
+
 @GetMapping("/nivelEducativo")
 	public String listStudents(Model model) {
 		model.addAttribute("nivelEducativo",nivelEducativoService.getAllNivelEducativo());
@@ -27,9 +55,21 @@ private NivelEducativoService nivelEducativoService;
 }
 @GetMapping("/nivelEducativo/new")
 public String createUsuarioForm(Model model){
-	NivelEducativo nivelEducativo = new NivelEducativo();
-	model.addAttribute("nivelEducativo", nivelEducativo);
-	return"nivelEducativo/create_nivelEducativo";
+	
+	try {
+		this.validarPerfil();
+		if(!this.perfil.getCVRol().equals("Consulta")) {
+			
+			NivelEducativo nivelEducativo = new NivelEducativo();
+			model.addAttribute("nivelEducativo", nivelEducativo);
+			return"nivelEducativo/create_nivelEducativo";
+		}else {
+			return "SinAcceso";
+		}
+		
+	}catch (Exception e) {
+		return "SinAcceso";
+	}
 }
 
 @PostMapping("/nivelEducativo")
@@ -40,13 +80,37 @@ public String saveNivelEducativo(@ModelAttribute("nivelEducativo") NivelEducativ
 
 @GetMapping("/nivelEducativo/{id}")
 public String deleteNivelEducativo(@PathVariable Integer id) {
-	nivelEducativoService.deleteNivelEducativoById(id);
-	return "redirect:/nivelEducativo";
+	
+	try {
+		this.validarPerfil();
+		if(!this.perfil.getCVRol().equals("Consulta")) {
+			
+			nivelEducativoService.deleteNivelEducativoById(id);
+			return "redirect:/nivelEducativo";
+		}else {
+			return "SinAcceso";
+		}
+		
+	}catch (Exception e) {
+		return "SinAcceso";
+	}
 }
 @GetMapping("/nivelEducativo/edit/{id}")
 public String editNivelEducativoForm(@PathVariable Integer id, Model model) {
-	model.addAttribute("nivelEducativo", nivelEducativoService.getNivelEducativoById(id));
-	return "nivelEducativo/edit_nivelEduactivo";
+	
+	try {
+		this.validarPerfil();
+		if(!this.perfil.getCVRol().equals("Consulta")) {
+			
+			model.addAttribute("nivelEducativo", nivelEducativoService.getNivelEducativoById(id));
+			return "nivelEducativo/edit_nivelEduactivo";
+		}else {
+			return "SinAcceso";
+		}
+		
+	}catch (Exception e) {
+		return "SinAcceso";
+	}
 }
 @PostMapping("/nivelEducativo/{id}")
 public String updateNivelEducativo (@PathVariable Integer id, @ModelAttribute ("nivelEducativo")NivelEducativo nivelEducativo, Model model) {
