+++ src/main/java/com/if7100/controller/OrganismoController.java
@@ -2,6 +2,8 @@ package com.if7100.controller;
 
 import java.util.Iterator;
 
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -9,30 +11,67 @@ import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 
+import com.if7100.entity.Hecho;
 import com.if7100.entity.Organismo;
+import com.if7100.entity.Perfil;
+import com.if7100.repository.UsuarioRepository;
 import com.if7100.service.OrganismoService;
+import com.if7100.service.PerfilService;
 
 @Controller
 public class OrganismoController {
 	
  private OrganismoService organismoService;
+//instancias para control de acceso
+ private UsuarioRepository usuarioRepository;
+ private Perfil perfil;
+ private PerfilService perfilService;
  
- public OrganismoController (OrganismoService organismoService) {
+ public OrganismoController (OrganismoService organismoService, PerfilService perfilService, UsuarioRepository usuarioRepository) {
 	 super();
 	 this.organismoService=organismoService;
+	 this.perfilService = perfilService;
+     this.usuarioRepository = usuarioRepository;
  }
+ 
+ private void validarPerfil() {
+ 	
+		try {
+			
+			Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+		    String username = authentication.getName();
+			
+			this.perfil = new Perfil(perfilService.getPerfilById(usuarioRepository.findByCVCedula(username).getCIPerfil()));
+			
+		}catch (Exception e) {
+			// TODO: handle exception
+		}
+		
+	}
  
  @GetMapping("/organismos")
  public String listStudents(Model model) {
 	 
 	 model.addAttribute("organismos",organismoService.getAllOrganismos());
-	 return "organismos";
+	 return "organismos/organismos";
  }
  
  @GetMapping("/organismos/new")
  public String createOrganismoForm(Model model) {
-	 model.addAttribute("organismo",new Organismo());
-	 return "create_organismo";
+	 
+	 try {
+			this.validarPerfil();
+			if(!this.perfil.getCVRol().equals("Consulta")) {
+				
+				model.addAttribute("organismo",new Organismo());
+				return "organismos/create_organismo";
+			}else {
+				return "SinAcceso";
+			}
+			
+		}catch (Exception e) {
+			return "SinAcceso";
+		}
  }
  
  @PostMapping("/organismos")
@@ -52,14 +91,37 @@ public class OrganismoController {
  @GetMapping("/organismos/{id}")
  public String deleteOrganismo(@PathVariable int id) {
 	 
-	 organismoService.deleteOrganismoById(id);
-	 return "redirect:/organismos";
+	 try {
+			this.validarPerfil();
+			if(!this.perfil.getCVRol().equals("Consulta")) {
+				
+				organismoService.deleteOrganismoById(id);
+				 return "redirect:/organismos";
+			}else {
+				return "SinAcceso";
+			}
+			
+		}catch (Exception e) {
+			return "SinAcceso";
+		}
  }
  
  @GetMapping("/organismos/edit/{id}")
  public String editOrganismoForm(Model model,@PathVariable int id) {
-	 model.addAttribute("organismo", organismoService.getOrganismoById(id));
-	 return "edit_organismo";
+	 
+	 try {
+			this.validarPerfil();
+			if(!this.perfil.getCVRol().equals("Consulta")) {
+				
+				model.addAttribute("organismo", organismoService.getOrganismoById(id));
+				return "organismos/edit_organismo";
+			}else {
+				return "SinAcceso";
+			}
+			
+		}catch (Exception e) {
+			return "SinAcceso";
+		}
  }
  
  @PostMapping("/organismos/{id}")
