+++ src/main/java/com/if7100/controller/OrientacionSexualController.java
@@ -1,7 +1,14 @@
 package com.if7100.controller;
 
+import com.if7100.entity.Hecho;
 import com.if7100.entity.OrientacionSexual;
+import com.if7100.entity.Perfil;
+import com.if7100.repository.UsuarioRepository;
 import com.if7100.service.OrientacionSexualService;
+import com.if7100.service.PerfilService;
+
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -13,23 +20,56 @@ import org.springframework.web.bind.annotation.PostMapping;
 public class OrientacionSexualController {
 	
  private OrientacionSexualService orientacionService;
+//instancias para control de acceso
+ private UsuarioRepository usuarioRepository;
+ private Perfil perfil;
+ private PerfilService perfilService;
  
- public OrientacionSexualController(OrientacionSexualService orientacionService) {
+ public OrientacionSexualController(OrientacionSexualService orientacionService, PerfilService perfilService, UsuarioRepository usuarioRepository) {
 	 super();
 	 this.orientacionService=orientacionService;
+	 this.perfilService = perfilService;
+     this.usuarioRepository = usuarioRepository;
  }
+ 
+ private void validarPerfil() {
+ 	
+		try {
+			
+			Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+		    String username = authentication.getName();
+			
+			this.perfil = new Perfil(perfilService.getPerfilById(usuarioRepository.findByCVCedula(username).getCIPerfil()));
+			
+		}catch (Exception e) {
+			// TODO: handle exception
+		}
+		
+	}
  
  @GetMapping("/orientacionessexuales")
  public String listOrientacionesSexuales(Model model) {
 	 
 	 model.addAttribute("orientacionesSexuales",orientacionService.getAllOrientacionesSexuales());
-	 return "orientacionesSexuales";
+	 return "orientacionesSexuales/orientacionesSexuales";
  }
  
  @GetMapping("/orientacionessexuales/new")
  public String createOrientacionSexualForm(Model model) {
-	 model.addAttribute("orientacionSexual",new OrientacionSexual());
-	 return "create_orientacionSexual";
+	 
+	 try {
+			this.validarPerfil();
+			if(!this.perfil.getCVRol().equals("Consulta")) {
+				
+				model.addAttribute("orientacionSexual",new OrientacionSexual());
+				return "orientacionesSexuales/create_orientacionSexual";
+			}else {
+				return "SinAcceso";
+			}
+			
+		}catch (Exception e) {
+			return "SinAcceso";
+		}
  }
  
  @PostMapping("/orientacionessexuales")
@@ -41,15 +81,37 @@ public class OrientacionSexualController {
  @GetMapping("/orientacionessexuales/{id}")
  public String deleteOrientacionSexual(@PathVariable int id) {
 	 
-	 orientacionService.deleteOrientacionSexualByCodigo(id);
-	 return "redirect:/orientacionessexuales";
+	 try {
+			this.validarPerfil();
+			if(!this.perfil.getCVRol().equals("Consulta")) {
+				
+				orientacionService.deleteOrientacionSexualByCodigo(id);
+				return "redirect:/orientacionessexuales";
+			}else {
+				return "SinAcceso";
+			}
+			
+		}catch (Exception e) {
+			return "SinAcceso";
+		}
  }
  
  @GetMapping("/orientacionessexuales/edit/{id}")
  public String editOrientacionForm(Model model,@PathVariable int id) {
 	 
-	 model.addAttribute("orientacionSexual", orientacionService.getOrientacionSexualByCodigo(id));
-	 return "edit_orientacionSexual";
+	 try {
+			this.validarPerfil();
+			if(!this.perfil.getCVRol().equals("Consulta")) {
+				
+				model.addAttribute("orientacionSexual", orientacionService.getOrientacionSexualByCodigo(id));
+				return "orientacionesSexuales/edit_orientacionSexual";
+			}else {
+				return "SinAcceso";
+			}
+			
+		}catch (Exception e) {
+			return "SinAcceso";
+		}
  }
  
  @PostMapping("/orientacionessexuales/{id}")
