+++ src/main/java/com/if7100/controller/RegistroPerfilController.java
@@ -3,6 +3,8 @@
  */
 package com.if7100.controller;
 
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -10,7 +12,9 @@ import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 
+import com.if7100.entity.Hecho;
 import com.if7100.entity.Perfil;
+import com.if7100.repository.UsuarioRepository;
 import com.if7100.service.PerfilService;
 
 
@@ -18,25 +22,66 @@ import com.if7100.service.PerfilService;
 public class RegistroPerfilController {
 
 	private PerfilService perfilService;
+	//instancias para control de acceso
+    private UsuarioRepository usuarioRepository;
+    private Perfil perfil;
 
-	public RegistroPerfilController(PerfilService perfilService) {
+	public RegistroPerfilController(PerfilService perfilService, UsuarioRepository usuarioRepository) {
 		super();
 		this.perfilService = perfilService;
+		this.usuarioRepository = usuarioRepository;
 	}
+	
+	 private void validarPerfil() {
+	    	
+			try {
+				
+				Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+			    String username = authentication.getName();
+				
+				this.perfil = new Perfil(perfilService.getPerfilById(usuarioRepository.findByCVCedula(username).getCIPerfil()));
+				
+			}catch (Exception e) {
+				// TODO: handle exception
+			}
+			
+		}
 	
 	@GetMapping("/perfiles")
 	public String listPerfiles(Model model) {
 		
-		model.addAttribute("perfiles", perfilService.getAllPerfiles());
-		return "perfiles";
+		try {
+			this.validarPerfil();
+			if(this.perfil.getCVRol().equals("Administrador")) {
+				
+				model.addAttribute("perfiles", perfilService.getAllPerfiles());
+				return "perfiles/perfiles";
+			}else {
+				return "SinAcceso";
+			}
+			
+		}catch (Exception e) {
+			return "SinAcceso";
+		}
 	}
 	
 	@GetMapping("/registroPerfil")
 	public String createPerfilForm (Model model) {
 		
-		Perfil perfil = new Perfil();
-		model.addAttribute("perfil", perfil);
-		return "registroPerfil";
+		try {
+			this.validarPerfil();
+			if(this.perfil.getCVRol().equals("Administrador")) {
+				
+				Perfil perfil = new Perfil();
+				model.addAttribute("perfil", perfil);
+				return "perfiles/registroPerfil";
+			}else {
+				return "SinAcceso";
+			}
+			
+		}catch (Exception e) {
+			return "SinAcceso";
+		}
 	}
 	
 	@PostMapping("/registroPerfil")
@@ -49,15 +94,37 @@ public class RegistroPerfilController {
 	@GetMapping("/perfiles/{id}")
 	public String deletePerfil (@PathVariable Integer id) {
 		
-		perfilService.deletePerfilById(id);
-		return "redirect:/perfiles";
+		try {
+			this.validarPerfil();
+			if(this.perfil.getCVRol().equals("Administrador")) {
+				
+				perfilService.deletePerfilById(id);
+				return "redirect:/perfiles";
+			}else {
+				return "SinAcceso";
+			}
+			
+		}catch (Exception e) {
+			return "SinAcceso";
+		}
 	}
 	
 	@GetMapping("/perfiles/edit/{id}")
 	public String editUsuarioForm (@PathVariable Integer id, Model model) {
 		
-		model.addAttribute("perfil", perfilService.getPerfilById(id));
-		return "edit_perfil";
+		try {
+			this.validarPerfil();
+			if(this.perfil.getCVRol().equals("Administrador")) {
+				
+				model.addAttribute("perfil", perfilService.getPerfilById(id));
+				return "perfiles/edit_perfil";
+			}else {
+				return "SinAcceso";
+			}
+			
+		}catch (Exception e) {
+			return "SinAcceso";
+		}
 	}
 	
 	@PostMapping("/perfiles/{id}")
